+++ packages/cli-module-build/src/lib/runner/startEmbeddedDb.ts
@@ -16,9 +16,14 @@
 
 import os from 'node:os';
 import fs from 'fs-extra';
-import { resolve as resolvePath } from 'node:path';
+import {
+  isAbsolute as isAbsolutePath,
+  resolve as resolvePath,
+} from 'node:path';
 import { getPortPromise } from 'portfinder';
 import { ForwardedError } from '@backstage/errors';
+import { ConfigSources } from '@backstage/config-loader';
+import { targetPaths } from '@backstage/cli-common';
 import chalk from 'chalk';
 
 const TEMP_DIR_PREFIX = 'backstage-dev-db-';
@@ -59,7 +64,38 @@ export interface EmbeddedDbConnectionConfig {
   password?: string;
 }
 
-export async function startEmbeddedDb(userConfig?: EmbeddedDbConnectionConfig) {
+export interface StartEmbeddedDbOptions {
+  configPaths?: string[];
+  targetDir?: string;
+}
+
+/**
+ * Reads the database configuration from the app config and, if the client is
+ * `embedded-postgres`, starts an embedded Postgres instance. Returns the config
+ * override to inject into the child process, or `undefined` if the database
+ * client is not `embedded-postgres`.
+ */
+export async function startEmbeddedDb(options: StartEmbeddedDbOptions): Promise<
+  | {
+      configOverride: Record<string, unknown>;
+      close(): Promise<void>;
+    }
+  | undefined
+> {
+  const dbConfig = await readDatabaseConfig(
+    options.configPaths,
+    options.targetDir,
+  );
+  if (dbConfig?.client !== 'embedded-postgres') {
+    return undefined;
+  }
+
+  return startEmbeddedDbInternal(dbConfig.connection);
+}
+
+async function startEmbeddedDbInternal(
+  userConfig?: EmbeddedDbConnectionConfig,
+) {
   console.warn(
     chalk.yellow(
       'WARNING: Using embedded-postgres for local development is experimental and subject to change',
@@ -107,6 +143,9 @@ export async function startEmbeddedDb(userConfig?: EmbeddedDbConnectionConfig) {
     throw error;
   }
 
+  const configOverride: Record<string, unknown> = {
+    client: 'pg',
+  };
   const defaultedConnection: Record<string, string | number> = {};
   if (!userConfig?.host) {
     defaultedConnection.host = host;
@@ -120,18 +159,71 @@ export async function startEmbeddedDb(userConfig?: EmbeddedDbConnectionConfig) {
   if (!userConfig?.port) {
     defaultedConnection.port = port;
   }
+  if (Object.keys(defaultedConnection).length > 0) {
+    configOverride.connection = defaultedConnection;
+  }
 
   return {
-    connection: {
-      host,
-      user,
-      password,
-      port,
-    },
-    defaultedConnection,
+    configOverride,
     async close() {
       await pg.stop();
       await fs.remove(tmpDir);
     },
   };
+}
+
+async function readDatabaseConfig(
+  configPaths?: string[],
+  targetDir?: string,
+): Promise<
+  | {
+      client: string;
+      connection?: EmbeddedDbConnectionConfig;
+    }
+  | undefined
+> {
+  const rootDir = targetPaths.rootDir;
+  const configBaseDir = targetDir ?? rootDir;
+  const source = ConfigSources.default({
+    rootDir,
+    allowMissingDefaultConfig: true,
+    argv: (configPaths ?? []).flatMap(p => [
+      '--config',
+      isAbsolutePath(p) ? p : resolvePath(configBaseDir, p),
+    ]),
+  });
+
+  const config = await ConfigSources.toConfig(source);
+  try {
+    const client = config.getOptionalString('backend.database.client');
+    if (!client) {
+      return undefined;
+    }
+
+    // Only read structured connection config if the value is an object;
+    // it can also be a plain string (e.g. ':memory:' for better-sqlite3).
+    const rawConnection = config.getOptional('backend.database.connection');
+    if (typeof rawConnection === 'string' || rawConnection === undefined) {
+      return { client };
+    }
+
+    const host = config.getOptionalString('backend.database.connection.host');
+    const port = config.getOptionalNumber('backend.database.connection.port');
+    const user = config.getOptionalString('backend.database.connection.user');
+    const password = config.getOptionalString(
+      'backend.database.connection.password',
+    );
+
+    const connection =
+      host !== undefined ||
+      port !== undefined ||
+      user !== undefined ||
+      password !== undefined
+        ? { host, port, user, password }
+        : undefined;
+
+    return { client, connection };
+  } finally {
+    config.close();
+  }
 }
